+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tournaments/TournamentDialogs.java
@@ -1,12 +1,12 @@
 package de.mephisto.vpin.ui.tournaments;
 
 import de.mephisto.vpin.connectors.mania.model.Tournament;
+import de.mephisto.vpin.connectors.mania.model.TournamentTable;
 import de.mephisto.vpin.ui.tournaments.dialogs.TournamentBrowserDialogController;
 import de.mephisto.vpin.ui.tournaments.dialogs.TournamentEditDialogController;
+import de.mephisto.vpin.ui.tournaments.dialogs.TournamentTableSelectorDialogController;
 import de.mephisto.vpin.ui.tournaments.view.TournamentTreeModel;
 import de.mephisto.vpin.ui.util.Dialogs;
-import de.mephisto.vpin.ui.vps.dialogs.VPSTableSelectorDialogController;
-import de.mephisto.vpin.ui.vps.containers.VpsSelection;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import javafx.scene.control.TreeItem;
 import javafx.stage.Stage;
@@ -16,7 +16,7 @@ public class TournamentDialogs {
   public static TreeItem<TournamentTreeModel> openTournamentDialog(@NonNull String title, @NonNull Tournament tournament) {
     Stage stage = Dialogs.createStudioDialogStage(TournamentEditDialogController.class, "dialog-tournament-edit.fxml", title);
     TournamentEditDialogController controller = (TournamentEditDialogController) stage.getUserData();
-    controller.setTournament(tournament);
+    controller.setTournament(stage, tournament);
     stage.showAndWait();
 
     return controller.getTournament();
@@ -30,12 +30,16 @@ public class TournamentDialogs {
     return controller.getTournament();
   }
 
-  public static VpsSelection openTableSelectionDialog(Stage parent) {
-    Stage stage = Dialogs.createStudioDialogStage(parent, VPSTableSelectorDialogController.class, "dialog-vps-table-selector.fxml", "Virtual Pinball Spreadsheet - Table Selection");
-    VPSTableSelectorDialogController controller = (VPSTableSelectorDialogController) stage.getUserData();
+  public static TournamentTable openTableSelectionDialog(Stage parent, Tournament tournament, TournamentTable tournamentTable) {
+    Stage stage = Dialogs.createStudioDialogStage(parent, TournamentTableSelectorDialogController.class, "dialog-tournament-table-selector.fxml", "Tournament - Table Selection");
+    TournamentTableSelectorDialogController controller = (TournamentTableSelectorDialogController) stage.getUserData();
+    if (tournamentTable != null) {
+      controller.setTournamentTable(tournament, tournamentTable);
+    }
+
     stage.showAndWait();
 
-    return controller.getSelection();
+    return controller.getTournamentTable();
   }
 
 }