+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tournaments/TournamentHelper.java
@@ -2,6 +2,8 @@ package de.mephisto.vpin.ui.tournaments;
 
 import de.mephisto.vpin.connectors.mania.model.Cabinet;
 import de.mephisto.vpin.connectors.mania.model.Tournament;
+import de.mephisto.vpin.connectors.mania.model.TournamentTable;
+import de.mephisto.vpin.restclient.games.GameRepresentation;
 
 public class TournamentHelper {
 
@@ -17,5 +19,20 @@ public class TournamentHelper {
     }
 
     return false;
+  }
+
+  public static String getIconColor(TournamentTable value) {
+    if (!value.isEnabled()) {
+      return "#B0ABAB";
+    }
+    return null;
+  }
+
+  public static String getLabelCss(TournamentTable value) {
+    String status = "";
+    if (!value.isEnabled()) {
+      status = "-fx-font-color: #B0ABAB;-fx-text-fill:#B0ABAB;";
+    }
+    return status;
   }
 }