+++ vpin-studio-ui/src/main/java/module-info.java
@@ -92,8 +92,6 @@ module de.mephisto.vpin.ui {
   opens de.mephisto.vpin.ui.players.dialogs to javafx.fxml;
   exports de.mephisto.vpin.ui.vps.containers;
   opens de.mephisto.vpin.ui.vps.containers to javafx.fxml;
-  exports de.mephisto.vpin.ui.vps.dialogs;
-  opens de.mephisto.vpin.ui.vps.dialogs to javafx.fxml;
   exports de.mephisto.vpin.ui.util.binding;
   opens de.mephisto.vpin.ui.util.binding to javafx.fxml;
   exports de.mephisto.vpin.ui.preferences.dialogs;